Integrate flatcontainer bfs in regular bfs
